----1---- Changed Carb/Protein/Fat in meals from int to float. Carb/Protein/Fat in user have remain unchanged. The total C/P/F for a meal are the base_macro * multiplier but rounded to the nearest int. Example: Before adding the Meal: Daily macros consumed where 10c, 20p, 10f Created a new Meal = Chicken, 0.0c, 33.3p, 4.7f, 100g Added the same Meal but with different amount = 45g, => multiplier = 45/100 = 0.45 Daily Meal entry = Chicken  0.0c, 33.3*0.45 = 14.985p => rounded to 15p, 4.7f*0.45 = 2.115 => rounded to 2f After adding the Meal: Daily macros consumed become 10+0 = 0c, 20 + 15 = 35p, 10 +2 = 12f ----2---- Changed serving from int to float. Made all neccessary GUI changes in order to make app functional (still need to revisit, re-evaluate and optimize them).

+++ app/src/main/java/com/example/kareem/macrotracker/Activities/ViewSavedMealsActivity.java
@@ -45,7 +45,6 @@ public class ViewSavedMealsActivity extends AppCompatActivity implements Adapter
 
         addmeal=(Button)findViewById(R.id.btn_addnewmeal);
         addmeal.setOnClickListener(this);
-
     }
 
     @Override
@@ -70,9 +69,9 @@ public class ViewSavedMealsActivity extends AppCompatActivity implements Adapter
                 int meal_id = C.getInt(0);      //meal)id
                 String meal_name = C.getString(1);   //meal_name
                 String date_created = C.getString(2);   //date_created
-                int carbs = C.getInt(3);      //carbs
-                int protein = C.getInt(4);      //protein
-                int fat = C.getInt(5);      //fat
+                float carbs = C.getFloat(3);      //carbs
+                float protein = C.getFloat(4);      //protein
+                float fat = C.getFloat(5);      //fat
                 portion = portion.values()[C.getInt(6)];    //portion
                 int daily_consumption = C.getInt(7);    //daily_consumption
                 int user_id = C.getInt(8);      //user_id
@@ -84,7 +83,6 @@ public class ViewSavedMealsActivity extends AppCompatActivity implements Adapter
 
     @Override
     protected void onPause() {
-
         super.onPause();
     }
